+++ tests/recvmsg.c
@@ -100,7 +100,7 @@ main(void)
 		w0_c, LENGTH_OF(w0_c),
 		w1_c, LENGTH_OF(w1_c),
 		w2_c, LENGTH_OF(w2_c),
-		ARRAY_SIZE(w_iov_), w_len,
+		(unsigned int) ARRAY_SIZE(w_iov_), w_len,
 		LENGTH_OF(w0_c), w0_d, w0_c,
 		LENGTH_OF(w1_c), w1_d, w1_c,
 		LENGTH_OF(w2_c), w2_d, w2_c);
@@ -127,7 +127,8 @@ main(void)
 		", msg_controllen=0, msg_flags=0}, 0) = %u\n"
 		" * %u bytes in buffer 0\n"
 		" | 00000 %-49s  %-16s |\n",
-		r0_c, r_len, ARRAY_SIZE(r0_iov_), r_len, r_len, r0_d, r0_c);
+		r0_c, r_len, (unsigned int) ARRAY_SIZE(r0_iov_),
+		r_len, r_len, r0_d, r0_c);
 
 	void *r1 = tail_alloc(r_len);
 	void *r2 = tail_alloc(w_len);
@@ -151,8 +152,8 @@ main(void)
 		", msg_iovlen=%u, msg_controllen=0, msg_flags=0}, 0) = %u\n"
 		" * %u bytes in buffer 0\n"
 		" | 00000 %-49s  %-16s |\n",
-		r1_c, r_len, w_len, ARRAY_SIZE(r1_iov_), w_len - r_len,
-		w_len - r_len, r1_d, r1_c);
+		r1_c, r_len, w_len, (unsigned int) ARRAY_SIZE(r1_iov_),
+		w_len - r_len, w_len - r_len, r1_d, r1_c);
 	close(0);
 
 	tprintf("+++ exited with 0 +++\n");
